fix(gate-client/gate): directory changes and renaming, unused dependencies and variables removal
