+++ bpdm-gate/pom.xml
@@ -41,7 +41,7 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>bpdm-gate-client</artifactId>
+            <artifactId>bpdm-gate-api</artifactId>
         </dependency>
         <dependency>
             <groupId>org.jetbrains.kotlin</groupId>
